Revert all changes to metagraph.py and subtensor.py, as Roman will be basing his changes on this PR.
