+++ bittensor/core/metagraph.py
@@ -1,4 +1,3 @@
-import asyncio
 import copy
 import os
 import pickle
@@ -24,8 +23,6 @@ from .chain_data import AxonInfo
 # For annotation purposes
 if typing.TYPE_CHECKING:
     from bittensor.core.subtensor import Subtensor
-    from bittensor.core.async_subtensor import AsyncSubtensor
-    from .chain_data import NeuronInfo, NeuronInfoLite
 
 
 METAGRAPH_STATE_DICT_NDARRAY_KEYS = [
@@ -69,25 +66,22 @@ This list defines the set of keys expected in the metagraph's state dictionary w
 """
 
 
-def get_save_dir(
-    network: str, netuid: int, root_dir: Optional[list[str]] = None
-) -> str:
+def get_save_dir(network: str, netuid: int) -> str:
     """
     Returns a directory path given ``network`` and ``netuid`` inputs.
 
     Args:
         network (str): Network name.
         netuid (int): Network UID.
-        root_dir: list to the file path for the root directory of your metagraph saves (i.e. ['/', 'tmp', 'metagraphs'],
-            defaults to ["~", ".bittensor", "metagraphs"]
 
     Returns:
         str: Directory path.
     """
-    _root_dir = root_dir or ["~", ".bittensor", "metagraphs"]
     return os.path.expanduser(
         os.path.join(
-            *_root_dir,
+            "~",
+            ".bittensor",
+            "metagraphs",
             f"network-{str(network)}",
             f"netuid-{str(netuid)}",
         )
@@ -223,7 +217,6 @@ class MetagraphMixin(ABC):
     axons: list[AxonInfo]
     chain_endpoint: Optional[str]
     subtensor: Optional["Subtensor"]
-    neurons: list[Union["NeuronInfo", "NeuronInfoLite"]]
 
     @property
     def S(self) -> Union[NDArray, "torch.nn.Parameter"]:
@@ -827,13 +820,9 @@ class MetagraphMixin(ABC):
             )
         return tensor_param
 
-    def save(self, root_dir: Optional[list[str]] = None) -> "Metagraph":
+    def save(self) -> "Metagraph":
         """
         Saves the current state of the metagraph to a file on disk. This function is crucial for persisting the current state of the network's metagraph, which can later be reloaded or analyzed. The save operation includes all neuron attributes and parameters, ensuring a complete snapshot of the metagraph's state.
-
-        Args:
-            root_dir: list to the file path for the root directory of your metagraph saves
-                (i.e. ['/', 'tmp', 'metagraphs'], defaults to ["~", ".bittensor", "metagraphs"]
 
         Returns:
             metagraph (bittensor.core.metagraph.Metagraph): The metagraph instance after saving its state.
@@ -853,7 +842,7 @@ class MetagraphMixin(ABC):
 
                 metagraph.load_from_path(dir_path)
         """
-        save_directory = get_save_dir(self.network, self.netuid, root_dir=root_dir)
+        save_directory = get_save_dir(self.network, self.netuid)
         os.makedirs(save_directory, exist_ok=True)
         if use_torch():
             graph_filename = f"{save_directory}/block-{self.block.item()}.pt"
@@ -869,7 +858,7 @@ class MetagraphMixin(ABC):
                 pickle.dump(state_dict, graph_file)
         return self
 
-    def load(self, root_dir: Optional[list[str]] = None) -> None:
+    def load(self):
         """
         Loads the state of the metagraph from the default save directory. This method is instrumental for restoring the metagraph to its last saved state. It automatically identifies the save directory based on the ``network`` and ``netuid`` properties of the metagraph, locates the latest block file in that directory, and loads all metagraph parameters from it.
 
@@ -878,10 +867,6 @@ class MetagraphMixin(ABC):
         the exact state it was in at the last save point, maintaining consistency in the network's representation.
 
         The method delegates to ``load_from_path``, supplying it with the directory path constructed from the metagraph's current ``network`` and ``netuid`` properties. This abstraction simplifies the process of loading the metagraph's state for the user, requiring no direct path specifications.
-
-        Args:
-            root_dir: list to the file path for the root directory of your metagraph saves
-                (i.e. ['/', 'tmp', 'metagraphs'], defaults to ["~", ".bittensor", "metagraphs"]
 
         Returns:
             metagraph (bittensor.core.metagraph.Metagraph): The metagraph instance after loading its state from the default directory.
@@ -896,7 +881,7 @@ class MetagraphMixin(ABC):
         Note:
             The default save directory is determined based on the metagraph's ``network`` and ``netuid`` attributes. It is important to ensure that these attributes are set correctly and that the default save directory contains the appropriate state files for the metagraph.
         """
-        self.load_from_path(get_save_dir(self.network, self.netuid, root_dir=root_dir))
+        self.load_from_path(get_save_dir(self.network, self.netuid))
 
     @abstractmethod
     def load_from_path(self, dir_path: str) -> "Metagraph":
@@ -1359,130 +1344,6 @@ class NonTorchMetagraph(MetagraphMixin):
         if "bonds" in state_dict:
             self.bonds = state_dict["bonds"]
         return self
-
-
-class AsyncMetagraph(NonTorchMetagraph):
-    """
-    AsyncMetagraph is only available for non-torch
-    """
-
-    def __init__(
-        self,
-        netuid: int,
-        network: str = settings.DEFAULT_NETWORK,
-        lite: bool = True,
-        sync: bool = True,
-        subtensor: "AsyncSubtensor" = None,
-    ):
-        if use_torch():
-            raise Exception("AsyncMetagraph is only available for non-torch")
-
-        # important that sync=False here bc the sync method of the superclass expects a sync Subtensor object,
-        # so will not work if run as True
-        NonTorchMetagraph.__init__(self, netuid, network, lite, False, subtensor)
-        if sync:
-            asyncio.get_running_loop().run_until_complete(
-                self.sync(block=None, lite=lite, subtensor=subtensor)
-            )
-
-    async def _initialize_subtensor(self, subtensor: "AsyncSubtensor"):
-        if subtensor and subtensor != self.subtensor:
-            self.subtensor = subtensor
-        if not subtensor and self.subtensor:
-            subtensor = self.subtensor
-        if not subtensor:
-            # Lazy import due to circular import (subtensor -> metagraph, metagraph -> subtensor)
-            from bittensor.core.async_subtensor import AsyncSubtensor
-
-            subtensor = AsyncSubtensor(network=self.chain_endpoint)
-            async with subtensor:
-                self.subtensor = subtensor
-        return subtensor
-
-    async def sync(
-        self,
-        block: Optional[int] = None,
-        lite: bool = True,
-        subtensor: Optional["AsyncSubtensor"] = None,
-    ):
-        # Initialize subtensor
-        subtensor = await self._initialize_subtensor(subtensor)
-        async with subtensor:
-            if (
-                subtensor.chain_endpoint != settings.ARCHIVE_ENTRYPOINT
-                or subtensor.network != "archive"
-            ):
-                cur_block = await subtensor.get_current_block()
-                if block and block < (cur_block - 300):
-                    logging.warning(
-                        "Attempting to sync longer than 300 blocks ago on a non-archive node. Please use the 'archive' "
-                        "network for subtensor and retry."
-                    )
-            block = block or await subtensor.block
-
-            # Assign neurons based on 'lite' flag
-            await self._assign_neurons(block, lite, subtensor)
-
-            # Set attributes for metagraph
-            self._set_metagraph_attributes(block, subtensor)
-
-            # If not a 'lite' version, compute and set weights and bonds for each neuron
-            if not lite:
-                await self._set_weights_and_bonds(subtensor=subtensor)
-
-    async def _assign_neurons(
-        self, block: int, lite: bool, subtensor: "AsyncSubtensor"
-    ):
-        if lite:
-            self.neurons = await subtensor.neurons_lite(block=block, netuid=self.netuid)
-        else:
-            self.neurons = await subtensor.neurons(block=block, netuid=self.netuid)
-        self.lite = lite
-
-    async def _set_weights_and_bonds(
-        self, subtensor: Optional["AsyncSubtensor"] = None
-    ):
-        # TODO: Check and test the computation of weights and bonds
-        if self.netuid == 0:
-            self.weights = await self._process_root_weights(
-                [neuron.weights for neuron in self.neurons],
-                "weights",
-                subtensor,
-            )
-        else:
-            self.weights = self._process_weights_or_bonds(
-                [neuron.weights for neuron in self.neurons], "weights"
-            )
-            self.bonds = self._process_weights_or_bonds(
-                [neuron.bonds for neuron in self.neurons], "bonds"
-            )
-
-    async def _process_root_weights(
-        self, data: list, attribute: str, subtensor: "AsyncSubtensor"
-    ) -> NDArray:
-        data_array = []
-        _n_subnets, subnets = await asyncio.gather(
-            subtensor.get_total_subnets(), subtensor.get_subnets()
-        )
-        n_subnets = _n_subnets or 0
-        for item in data:
-            if len(item) == 0:
-                data_array.append(np.zeros(n_subnets, dtype=np.float32))
-            else:
-                uids, values = zip(*item)
-                data_array.append(
-                    convert_root_weight_uids_and_vals_to_tensor(
-                        n_subnets, list(uids), list(values), subnets
-                    )
-                )
-        tensor_param: NDArray = (
-            np.stack(data_array) if len(data_array) else np.array([], dtype=np.float32)
-        )
-        if len(data_array) == 0:
-            logging.warning(
-                f"Empty {attribute}_array on metagraph.sync(). The '{attribute}' tensor is empty."
-            )
-        return tensor_param
 
 
 Metagraph = TorchMetaGraph if use_torch() else NonTorchMetagraph
